refactored the EventList comp
